requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
python-decouple==3.7
django-cors-headers==4.3.0
requests==2.31.0
pandas
implicit
scikit-learn